pyproject.toml
[project]
name = "beancount-tools"
version = "0.1.0"
description = "Beancount importers and transaction processing tools for Chinese financial institutions"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beancount>=3.2.0",
    "beanquery>=0.2.0",
    "dateparser>=1.3.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
beancount-import = "beancount_tools.cli.import_cli:main"
beancount-postprocess = "beancount_tools.cli.postprocess:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/beancount_tools"]
